Update to Debian stretch libraries.
Note that Checkstyle and Cobertura are excluded from this update,
because they're not required for the build and because we're still
using newer versions than what's in stretch. We could attempt to
downgrade and see if the versions in stretch are new enough. But
that's not part of this patch.