Update PostgreSQL JDBC driver version.
There's an issue with Java 9+ complaining about an ilegal reflective
access [...] to field java.util.TimeZone.defaultTimeZone which is
fixed in PostgreSQL JDBC driver version 42.2.0. Updating to 42.2.5
which happens to be in Debian buster under the assumption that is has
seen more usage than 42.2.0 (containing the fix) or 42.2.7 (latest at
the moment).

Found while running integration tests using metrics-test.